market_radar/requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
slack-sdk>=3.21.0
pytz>=2023.3 